Player Implementation
Created inheritance hierarchy for Players with HumanPlayer and ComputerPlayer as subclasses of the abstract class Player.
Added methods for choosing the category for a round.
Added methods for returning a player's "name"
Added Card method to return ctageory with highest value for a card (used by the AI).
Added static String array to Card class to hold the names of all categories.